+++ src/main/java/module-info.java
@@ -1,11 +1,11 @@
 import org.cryptomator.integrations.autostart.AutoStartProvider;
-import org.cryptomator.integrations.sidebar.SidebarService;
 import org.cryptomator.integrations.keychain.KeychainAccessProvider;
+import org.cryptomator.integrations.quickaccess.QuickAccessService;
 import org.cryptomator.integrations.revealpath.RevealPathService;
 import org.cryptomator.integrations.uiappearance.UiAppearanceProvider;
 import org.cryptomator.windows.autostart.WindowsAutoStart;
-import org.cryptomator.windows.sidebar.ExplorerSidebarService;
 import org.cryptomator.windows.keychain.WindowsProtectedKeychainAccess;
+import org.cryptomator.windows.quickaccess.ExplorerQuickAccessService;
 import org.cryptomator.windows.revealpath.ExplorerRevealPathService;
 import org.cryptomator.windows.uiappearance.WinUiAppearanceProvider;
 
@@ -21,5 +21,5 @@ module org.cryptomator.integrations.win {
 	provides KeychainAccessProvider with WindowsProtectedKeychainAccess;
 	provides UiAppearanceProvider with WinUiAppearanceProvider;
 	provides RevealPathService with ExplorerRevealPathService;
-	provides SidebarService with ExplorerSidebarService;
+	provides QuickAccessService with ExplorerQuickAccessService;
 }
