+++ src/main/resources/META-INF/services/org.cryptomator.integrations.quickaccess.QuickAccessService
@@ -1,0 +1,1 @@
+org.cryptomator.windows.quickaccess.ExplorerQuickAccessService
