--- src/main/resources/META-INF/services/org.cryptomator.integrations.sidebar.SidebarService
@@ -1,1 +1,0 @@
-org.cryptomator.windows.sidebar.ExplorerSidebarService
