+++ src/test/java/org/cryptomator/windows/quickaccess/ExplorerQuickAccessServiceIT.java
@@ -1,7 +1,6 @@
-package org.cryptomator.filemanagersidebar;
+package org.cryptomator.windows.quickaccess;
 
-import org.cryptomator.integrations.sidebar.SidebarServiceException;
-import org.cryptomator.windows.sidebar.ExplorerSidebarService;
+import org.cryptomator.integrations.quickaccess.QuickAccessServiceException;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,7 +11,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
 
-public class ExplorerSidebarServiceIT {
+public class ExplorerQuickAccessServiceIT {
 
 	@TempDir
 	Path base;
@@ -20,13 +19,13 @@ public class ExplorerSidebarServiceIT {
 	@Test
 	@DisplayName("Integrates a temp dir for 20s into the file explorer sidebar")
 	@Disabled
-	public void testExplorerSidebarIntegration() throws IOException, InterruptedException, SidebarServiceException {
+	public void testExplorerSidebarIntegration() throws IOException, InterruptedException, QuickAccessServiceException {
 		var p = base.resolve("integration-win-testVault");
 		Files.createDirectory(p);
 		Files.createFile(p.resolve("firstLevel.file"));
 		var sub = Files.createDirectory(p.resolve("subdir"));
 		Files.createFile(sub.resolve("secondLevel.file"));
-		var entry = new ExplorerSidebarService().add(p, "integration-win-tempDir");
+		var entry = new ExplorerQuickAccessService().add(p, "integration-win-tempDir");
 
 		Thread.sleep(Duration.ofSeconds(20));
